Remove Collect and WebUtils
