+++ app/src/org/odk/collect/android/activities/DrawActivity.java
@@ -41,7 +41,7 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 
 import org.commcare.dalvik.R;
-import org.odk.collect.android.application.Collect;
+import org.odk.collect.android.application.ODKStorage;
 import org.odk.collect.android.utilities.FileUtils;
 
 import java.io.File;
@@ -108,9 +108,9 @@ public class DrawActivity extends Activity {
         if (extras == null) {
             loadOption = OPTION_DRAW;
             refImage = null;
-            savepointImage = new File(Collect.TMPDRAWFILE_PATH);
+            savepointImage = new File(ODKStorage.TMPDRAWFILE_PATH);
             savepointImage.delete();
-            output = new File(Collect.TMPFILE_PATH);
+            output = new File(ODKStorage.TMPFILE_PATH);
         } else {
             loadOption = extras.getString(OPTION);
             if (loadOption == null) {
@@ -129,7 +129,7 @@ public class DrawActivity extends Activity {
                     FileUtils.copyFile(refImage, savepointImage);
                 }
             } else {
-                savepointImage = new File(Collect.TMPDRAWFILE_PATH);
+                savepointImage = new File(ODKStorage.TMPDRAWFILE_PATH);
                 savepointImage.delete();
                 if (refImage != null && refImage.exists()) {
                     FileUtils.copyFile(refImage, savepointImage);
@@ -140,7 +140,7 @@ public class DrawActivity extends Activity {
             if (uri != null) {
                 output = new File(uri.getPath());
             } else {
-                output = new File(Collect.TMPFILE_PATH);
+                output = new File(ODKStorage.TMPFILE_PATH);
             }
         }
 
@@ -360,7 +360,7 @@ public class DrawActivity extends Activity {
             mBitmapPaint = new Paint(Paint.DITHER_FLAG);
             mCurrentPath = new Path();
             setBackgroundColor(0xFFFFFFFF);
-            mBackgroundBitmapFile = new File(Collect.TMPDRAWFILE_PATH);
+            mBackgroundBitmapFile = new File(ODKStorage.TMPDRAWFILE_PATH);
         }
 
         public DrawView(Context c, boolean isSignature, File f) {
